moved mainCards to mainpage.tsx
