+++ cpp-website/src/globals/CPPInfo.ts
@@ -1,5 +1,3 @@
-import { faPeopleGroup, faTrophy, faChalkboard, faCalendar } from "@fortawesome/free-solid-svg-icons";
-
 const CPPTheme = {
   lightblue: "#649bd3",
   blue: "#00599d",
@@ -18,39 +16,4 @@ const CPPCalendarConfig = {
   apiKey: "AIzaSyAASW9GZYCIxKLVjS8wEJ1tP-oNW3028eE",
 }
 
-const CPPMainCards = [
-  {
-    title: "Collaboration",
-    desc: `
-        We provide sessions for members to help each other progress
-        through the toughest problems they are confronted with.
-      `,
-    icon: faPeopleGroup
-  },
-  {
-    title: "Lectures",
-    desc: `
-        Lectures are ran by experienced competitive programmers
-        where we dive into challenging computer science topics.
-      `,
-    icon: faChalkboard
-  },
-  {
-    title: "Contests",
-    desc: `
-        We host practice contests where beginning programmers
-        to top competitors can compete at differently levels.
-      `,
-    icon: faTrophy
-  },
-  {
-    title: "Events",
-    desc: `
-        Seminars, conferences, and meetups are scheduled to 
-        interact with the broader computer science.
-      `,
-    icon: faCalendar
-  }
-]
-
-export {CPPTheme, email, githubOrg, discordInvite, logo, CPPCalendarConfig, CPPMainCards};
+export {CPPTheme, email, githubOrg, discordInvite, logo, CPPCalendarConfig};
